moved logout function from index to profile page and removed onAuthStateChanged function from profile screen as it was already used in main screen so no need to call it multiple times
